requirements.txt
langchain==0.1.17
pandas==2.2.2
faiss-cpu==1.8.0
python-box==7.1.1
ipykernel==6.29.4
-e .